Fix generate/import keypair example in API doc
fixes bug 1020614

As per the bug report, the example in the keypairs extension
was incorrect. If the optional public key is specified (i.e. a
keypair import), a private key will not be returned.

The most common case is to specify only a name and get back a
full keypair, so this patch modifies the examples to portray this.

Change-Id: I9c8f2269ee5374ef609a91bbc938066363a033cd