Modify API ref for per-project-user quotas CRUD methods
**tenant** is an entity that owns the resources,
sometines called a project, sometimes the same as
a user

In addition to the owning entity (tenant), openstack
stores the entity performing the action as the **user**.

1st tenant ID in the URI, is the ID of the admin user
2ed tenant ID in the URI, is the project ID/Name
User ID is specified vis query string.
(I tried my best to express what I mean using the
current WADL markup.)

Changed the url path structure.
http://api.openstack.org/api-ref-compute.html#os-quota-sets

Changed the sample responses to point to the new ones in
/api_samples/os-user-quotas/

For the other part, in the original bug description:
Add commands for project user quotas management.
     - Show user quotas:
       nova-manage project quota --project <Project name>
       --user <User name>
     - Update/Create user quotas:
       nova-manage project quota --project <Project name>
       --user <User name> --key <key> --value <value>
I found this page:
https://wiki.openstack.org/wiki/NovaManage
"The nova-manage isn't properly documented, but it's going away
in Folsom, so using this wiki page for quick docs on it."

Looks like the usage of this method still needs further
validation, so dose the API call.
It may need to be futher documented in
docs.openstack.org/trunk/openstack-ops/content/quotas.html

In the response:
"fixed_ips": -1
Does not seem to be right to have negative value for number
of IPs.
docs.openstack.org/user-guide-admin/content/cli_set_quotas.html
"Number of fixed IP addresses allowed per tenant. This number
must be equal to or greater than the number of allowed
instances."

Change-Id: Id4e08170ebf04ebb4ba846bf8b576a228b91441b
Closes-Bug: #1187277