Update object-api wadl for completeness and accuracy
The point of this change is to ensure that
these two docs are in sync and are complete with regards
to the descriptions of each method and their request
and response parameters:
http://api.openstack.org/api-ref-objectstorage.html
  (API Reference page for Object Storage)
http://docs.openstack.org/api/openstack-object-storage/1.0/content/
  (API Reference (or spec) for Object Storage)
In addition, the current documents don’t describe ACLs, FormPOST, TempURLs,
StaticWeb, Bulk Upload, Bulk Delete, the OPTIONS operation
(though this is implicit in CORS support) and maybe a couple of other things.
These items should be part of the https://review.openstack.org/#/c/64579/
review, which I'm still working on.
updated to add {name} for variables in metadata

Co-Author: Donagh McCabe

Partial-Bug: #1187119
Partial-Bug: #1214139

Change-Id: I82331b1c3cfc5b19f91f7ac1379476e6a9e6806d
author: diane fleming