Fix incorrect request format for rescoping a token
The request format for rescoping a token indicated that the
"token" was a string. It's actually an object where the token
id is in the "id" attribute.

The fix is to change the documentation so that it's more obvious
that the "token" is an object that contains an "id" attribute.

This will make it easier for users because they won't try to pass
the token as a string rather than in the "id" attribute.

Change-Id: I2865caa73418e6b26e3ac17686c430b910db4444
Closes-Bug: #1292844