Add VPNaaS Reference API Documentation
This commits restores the VPNaaS reference API documentation that used
to reside on the site, before conversion to WADL. The information for
this commit comes from the openstack-attic/netconn-api repository.

Includes edits for consistency and clarity (df).

The status of the API is set to CURRENT, as the API has not changed in
a long time and no changes are expected.

Change-Id: I2b4ac312aa62bb6081588d72f4df9882dfd6bde1
Closes-Bug: #1433561