Clarify XML 'deprecation' for Compute API
Effective with the Kilo release of the Compute API, support for the XML
content type in the bodies of requests and responses has been eliminated
for both v2 (supported) and v2.1 (current).  This commit changes the
API Reference note stating that XML is deprecated to indicate that
support has been completely eliminated.

Change-Id: I812c2467d5bab0d8a12f5fdd1653a8ce90723460
Closes-bug: #1474868