Fix dedent for C# code, upgrade sphinx
The C# include for the .NET SDK has to be indented to follow
the standards. This patch fixes up the includes so that the
indenting is appropriate for in-line reading of snippets

Since Sphinx 1.3 is required for dedent, this changes the
test requirements for the api-site repository to upgrade
the sphinx version.

Change-Id: I539df7bc2c08ac6b31ba8ba4b7b94cf66b3d58a4
Co-Authored-By: Liang Bo <liang.bo@99cloud.net>