Merge API doc of server rebuild.
The API doc of server rebuild and server rebuild with
preserve ephemeral option is seperate. As preserve
ephemeral is an optional parameter when doing server
rebuild, it will be better to merge these two together
and point out the meaning of preserve ephemeral option
in the description.

(df) Added parameter descriptions.

Change-Id: Idaced7d12a1833473ec42c784b056882f9be7d17
Closes-bug: #1505929