Rework introduction to Compute API
This reworks the introduction to the Compute API docs to include a link
to the Compute API concept guide that is currently being written.

It also removes the incomplete content page style list, and moves the
note about the dropping of XML support so it is not the first line of
the introduction.

Change-Id: I4ba1477e2ee1a66f11d565f08fabb97b5e9400c9