Fix the link of "create secgroup default rule"
The links of "create security group default rule" and "create
security group rule" are duplicated, and we cannot open the link
of "create security group rule" on api-site, because "create security
group default rule" is always open due to the order.
This patch fixes this problem with adding "Default" to the link for
matching the other "secgroup default rule" links.

Closes-Bug: #1521826

Change-Id: I9cdaf2a352a5b1d4267cc18309da02b83cb8e72a