Fixed default for networks on server creation
When creating a server, the user may pass a "networks" parameter.
The documentation said that the default is to attach the server to all
tenant's networks, but this is not true.
The default is to attach the server to the tenant network if there's
only one network in the tenant, but if there is more than one network,
the user must specify a network or an error will bethrown.

Change-Id: I656f1f272777e93f66d2e3717a76e64ae53b7a99
Closes-Bug: #1548815