[Orchestration API v1] Fix stack-adopt response
The response to adopting a stack is not the same as the response to
abandoning a stack. It is the same as the response to creating a stack. In
fact, it could not be otherwise because adopting a stack *is* creating a
stack and uses the exact same endpoint and the exact same code.

Change-Id: Ifa62c953fd968fdf4b536935b28ceedd8a56aaea