[Data Processing API] Missing information in Response Parameter tables.
When listing, creating, showing and updating jobs using Data Processing API v1.1,
the Response body of all requests contain the "interface" field in the information
of a job. However, in the Response Parameters table, the information
of "interface" parameter IS MISSED.
This patch add "interface" parameter to Response Parameters tables in the description
of requests.

Change-Id: I3c30a9e9a83185bf971ce35f9832847777882455
Closes-Bug: #1553105