Add API doc for misc trove operations
This  commit adds the following items that were missing
from the API doc for trove:

trove root-disable
trove detach-replica

trove create -- add the following missing args:
type
replica_of
replica_count
datastore
datastore_version
configuration

***

It also changes the following existing doc
for the following reasons:

in dbaas.wadl and ch_database-v1.xml-->

listConfigDefaults -- really belongs under a
separate "configuration" resource.

Because otherwise, it has the same signature as
"show instance details," which is not right.

***

in common.ent -->

databaseRequestParameter

The databases arg is optional when you create
an instance -- so required should be false.

datastoreNameTemplateParameter

Description used to read "The name of the data
store for which to list versions."

But now this param is used to get other stuff than
versions, so it needs to be more general, like
this:

"The name of the data store."

Change-Id: I24c7e772cd31f4362a277c54e63e9ced34b46de8
Partial-Bug: #1376972