Images api-ref: Redirect WADL docs to in-tree RST docs
The new RST-based, in-tree Images API Reference docs are ready
to take over for the old WADL-based docs.

The Images API Reference has been slightly restructured to pull
out the common /versions call and the /v2/metadefs API calls.
Thus the redirect for both the old v1 reference and the old
v2 reference redirect to the same place, that is, the "Image Service
APIs" landing page.

Change-Id: Iaf8ccbe1248b12e9e8fef48d52aeb7688fbac8a1