Drop DocBook support, swagger building
We do not need to build and publish the DocBook XML content anymore,
everything is frozen and in the process of moving to projects. Remove
set up for building of DocBook content.

This leaves the XML files still in tree, we can remove them later.

The DocBook XML gates are still setup, so some tox environments have
been made a noop.

Require openstack-doc-tools 1.0.0 which removes DocBook XML support.

Also, drop swagger building, this is not needed anymore.

Change-Id: I276953013a518a09dabb89bb8f80ef7739b4491b