Mark Cinder v1 as DEPRECATED
The Cinder v1 API has been deprecated for a while now[1][2][3], but
the api-quick-start index still shows it as SUPPORTED.  This patch
remedies that problem by moving it to the appropriate place.
A separate change will be submitted to the Cinder project to
correct the devref header as well.

[1] https://wiki.openstack.org/wiki/ReleaseNotes/Juno#Cinder_Features
[2] http://developer.openstack.org/api-ref-blockstorage-v1.html
[3] https://review.openstack.org/#/c/309652/

Change-Id: I3493c1440882c2d61442008ae1c5239cc42f3187
Partial-Bug: #1623642