QuickStart guide changes
Quick start guide currently uses the deprecated Identity API.
The old and new APIs use different request and response structures.
With this change, the guide now uses the current(3.x) identity API
to get a token.

Additionally, this change updates the file to use the term "project"
in place of the older term "tenant".

Change-Id: I4357c769af4ffcca56556dba8b11540c57145774
Closes-Bug: #1681121