don't duplicate the list of API reference guides
We have an automatically managed list of API reference guides in the
openstack-manuals repo being published to doc.openstack.org/api. The
results are validated against the list of actually published guides,
and will be kept up to date when projects come and go. It will also be
maintained as the API documentation moves into the main doc build for
each project as part of the documentation migration project [1].

This patch removes the redundant information from this repo and
replaces it with a link to the other repository.

[1] http://specs.openstack.org/openstack/docs-specs/specs/pike/os-manuals-migration.html

Change-Id: I80d7a6ed2ad8ccf4e505da1932b8a900cf4fe3f8
Signed-off-by: Doug Hellmann <doug@doughellmann.com>