Generate api-ref redirects for service type aliases
We track official service type names and their historical aliases. As
api-ref is organized by service-type, it stands to reason that if a
service-type changes (or has changed) that a link might exist to the old
value.

Use the service types authority data to generate a set of redirects for
all aliases of a service to the official service type location.

Change-Id: I96a5502ca6cd3e78f996ad9b60101c907e2150cd